Part6.a: Redux notes
Details:
---
* Use `react-redux` hooks to provide the Redux store in the `<App />`
  component
