+++ part6/redux-notes/src/reducers/noteReducer.js
@@ -16,4 +16,24 @@ const noteReducer = (state = [], action) => {
     }
 }
 
+const generateId = () => Number((Math.random() * 100000000).toFixed(0))
+
+export const createNote = (content) => {
+    return {
+        type: 'NEW_NOTE',
+        data: {
+            content,
+            important: false,
+            id: generateId(),
+        },
+    }
+}
+
+export const toggleImportanceOf = (id) => {
+    return {
+        type: 'TOGGLE_IMPORTANCE',
+        data: { id },
+    }
+}
+
 export default noteReducer
